Move Auth0 into src/js/context
