+++ src/js/pages/auth/Logout.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { useAuth0 } from 'js/services/react-auth0-spa';
+import { useAuth0 } from 'js/context/Auth0';
 import Simple from 'js/pages/templates/Simple';
 
 export default () => {
